Scrum-123: Nuxt FE(refactor): Major refactor to depend on the toast to  make request
Initially it was set up that as soon as you confirm/swipe (based on settings) the request will be made. If the user clicks the "undo", another request will be made to amend the previous action.

Now the request is only made if the toast disappears which means that the user can no longer undo his action.