SCRUM-233: Nuxt FE(feature): Make both logos redirect to index page
Middleware will take care of going to the dashboard from there. To be honest currently this feature isnt very useful since the pages that can have the redirect functionality dont usually have the logo visible in them.

but in can be used in the future if we expand our pages or layout