SCRUM-238: Nuxt FE(feature): replace the history stack instead of pushing a new one
When we use navigateTo, it pushes a new entry to the stack, so when we click back it goes to the last available item in the stack.

Instead we should replace the current stack item to the new page and replace it back to the profile. this way the before-last page is the one that came into the profile